chore: Code cleanup
