+++ src/SourceGenerators/Uno.UI.SourceGenerators/XamlGenerator/NameScope.cs
@@ -54,7 +54,7 @@ namespace Uno.UI.SourceGenerators.XamlGenerator
 
 		public List<string> XBindTryGetMethodDeclarations { get; } = [];
 
-		public List<string> ExplicitApplyMethods { get; } = [];
+		public List<string> ExplicitApplyMethods { get; } = []; // TODO: Merge with CallbackMethods
 
 		public List<string> ExplicitApplyMethodNames { get; } = [];
