+++ src/SourceGenerators/Uno.UI.SourceGenerators/XamlGenerator/XamlObjectDefinition.cs
@@ -4,8 +4,8 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
+using Uno.UI.SourceGenerators.XamlGenerator.Utils;
 using Uno.UI.SourceGenerators.XamlGenerator.XamlRedirection;
 
 namespace Uno.UI.SourceGenerators.XamlGenerator
@@ -74,51 +74,6 @@ namespace Uno.UI.SourceGenerators.XamlGenerator
 			return $"{owner.Key}_ø_{NamingHelper.GetShortName(Type.Name)}"; // Should not happen
 
 			
-		}
-	}
-
-	public class NamingHelper
-	{
-		/// <summary>
-		/// Gets a short name for the given XAML type name, eg.
-		/// * ListView => LiVi
-		/// * MyLongControlName123 => MyLoCoNa123
-		/// </summary>
-		/// <param name="name"></param>
-		/// <returns></returns>
-		public static string GetShortName(string name)
-		{
-			var sb = new StringBuilder();
-			var i = 0;
-			var len = name?.Length ?? 0;
-			var allowLower = true;
-
-			while (i < len)
-			{
-				if (char.IsLetterOrDigit(name, i))
-				{
-					var c = name![i];
-					if (char.IsUpper(c) || char.IsDigit(c))
-					{
-						allowLower = true;
-						sb.Append(c);
-					}
-					else if (allowLower)
-					{
-						allowLower = false;
-						sb.Append(c);
-					}
-				}
-				else
-				{
-					allowLower = true; // We ignore the char, but if's an _ we allow lower next time
-					i++;
-				}
-
-				i++;
-			}
-
-			return sb.ToString();
 		}
 	}
 }
